pyproject.toml
[tool.poetry]
name = "pi-blur"
version = "0.1.0"
description = ""
authors = ["Chris Barry <chris@greatjones.co>"]

[tool.poetry.scripts]
start = "pi_blur.__main__:main"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.25.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
